Added map6.txt correct file

+++ patients-transport-center/src/main/java/pl/group2/optimizer/impl/items/paths/Path.java
@@ -17,7 +17,7 @@ public class Path {
 
     public int getId() {return id; }
 
-    public double getDistance() {
+    public int getDistance() {
         return distance;
     }
